cleanup in order to reduce AbstractSingleActor size
